+++ generators/app/templates/app-vue-vuetify-ts/src/plugins/vuetify.ts
@@ -1,0 +1,11 @@
+import Vue from 'vue';
+// @ts-ignore
+import Vuetify from 'vuetify/lib';
+
+Vue.use(Vuetify);
+
+export default new Vuetify({
+    icons: {
+        iconfont: 'fa',
+    },
+});